+++ src/mimeparser.rs
@@ -216,9 +216,16 @@ impl MimeMessage {
         headers.remove("secure-join-fingerprint");
         headers.remove("chat-verified");
 
+        let is_thunderbird = if let Some(user_agent) = headers.get("user-agent") {
+            info!(context, "Detected thunderbird");
+            user_agent.contains("Thunderbird")
+        } else {
+            false
+        };
+
         let from = from.context("No from in message")?;
         let mut decryption_info =
-            prepare_decryption(context, &mail, &from.addr, message_time).await?;
+            prepare_decryption(context, &mail, &from.addr, message_time, is_thunderbird).await?;
 
         // Memory location for a possible decrypted message.
         let mut mail_raw = Vec::new();